+++ totalRP3/modules/flyway/flyway.lua
@@ -21,7 +21,7 @@ TRP3_API.flyway = {};
 
 local type, tostring = type, tostring;
 
-local SCHEMA_VERSION = 6;
+local SCHEMA_VERSION = 7;
 
 if not TRP3_Flyway then
 	TRP3_Flyway = {};